requirements.txt
torch>=2.6.0
torchvision>=0.21.0
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
seaborn>=0.13.0
Pillow>=10.0.0
opencv-python>=4.8.0
optuna>=3.5.0
customtkinter>=5.2.0
tkinterdnd2>=0.4.0
